Healing + bug fix
bug with status update fixed, time instead of ticks in AI,
bug with dead enemyes under AI control fixed, healing create

+++ AIController.h
@@ -47,7 +47,7 @@ namespace MagicWars_NS {
         TouchControl& d_touchControl;
         std::map<std::string, AIEngine*> d_engines;
         int d_stage = 0;
-        int d_timer = 0;
+        float d_timer = 0;
     };
 }
